PipeOps and HandlerOps as implicit classes
